Fixed Makefile and created user_conf template
Makefile will not be fixed to a java version using JAVA_HOME environment variable to get the C headers to compile the native code. Also, create src/main/resources/RENAMEMEuser_conf file to be renamed and changed its content, so the tests can run fine if a valid user is specified and cgroups is installed and started. Probably these and other informations (such as the Cgroups installation/check if it is working/etc) needs to be added into the README.md